Fixed UP pages for the rest of the site
